IMPLEMENTADNO COMPARATIVO EM DASHBOARD

+++ app/api/BI/schemas/departamento_schema.py
@@ -1,7 +1,6 @@
 from pydantic import BaseModel
-from decimal import Decimal
 
 
 class VendasPorDepartamento(BaseModel):
     departamento: str  # agora é nome
-    total_vendas: Decimal
+    total_vendas: float
